+++ src/main/java/com/sumin/paypay/PayPayApplication.java
@@ -1,0 +1,13 @@
+package com.sumin.paypay;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PayPayApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PayPayApplication.class, args);
+	}
+
+}
