+++ src/main/java/com/sumin/paypay/entity/User.java
@@ -1,0 +1,48 @@
+package com.sumin.paypay.entity;
+
+import java.time.Instant;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@Entity
+@Builder
+@Table(name = "user")
+@AllArgsConstructor
+public class User {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "id")
+	private Long id;
+
+	@Column(name = "name", nullable = false)
+	private String name;
+
+	@Column(name = "email", nullable = false)
+	private String email;
+
+	@Column(name = "pwd", nullable = false)
+	private String pwd;
+
+	@Column(name = "active", nullable = false)
+	private Boolean active;
+
+	@Column(name = "created_at", nullable = false)
+	private Instant createdAt;
+
+	@Column(name = "updated_at", nullable = false)
+	private Instant updatedAt;
+
+
+}
