+++ src/main/java/com/sumin/paypay/exception/CommonException.java
@@ -1,0 +1,12 @@
+package com.sumin.paypay.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class CommonException extends RuntimeException{
+
+	private final ResultCode resultCode;
+	private final String message;
+}
