+++ src/main/java/com/sumin/paypay/exception/ErrorResponse.java
@@ -1,0 +1,10 @@
+package com.sumin.paypay.exception;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ErrorResponse {
+	private final String message;
+}
