+++ src/main/java/com/sumin/paypay/exception/ResultCode.java
@@ -1,0 +1,10 @@
+package com.sumin.paypay.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface ResultCode {
+
+	HttpStatus getStatus();
+
+	String getMessage();
+}
