+++ src/main/java/com/sumin/paypay/model/AuthRequest.java
@@ -1,0 +1,8 @@
+package com.sumin.paypay.model;
+
+import jakarta.validation.constraints.NotNull;
+
+public interface AuthRequest {
+	@NotNull
+	Long getUserId();
+}
