+++ src/main/java/com/sumin/paypay/model/request/AccountCreateRequest.java
@@ -1,0 +1,19 @@
+package com.sumin.paypay.model.request;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class AccountCreateRequest {
+
+	@NotNull
+	private Long userId;
+
+}
