+++ src/main/java/com/sumin/paypay/model/request/AccountDeleteRequest.java
@@ -1,0 +1,18 @@
+package com.sumin.paypay.model.request;
+
+import com.sumin.paypay.model.AuthRequest;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class AccountDeleteRequest implements AuthRequest {
+
+	private final Long userId;
+
+	@Override
+	public Long getUserId() {
+		return null;
+	}
+}
