+++ src/main/java/com/sumin/paypay/model/response/AccountCreateResponse.java
@@ -1,0 +1,10 @@
+package com.sumin.paypay.model.response;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class AccountCreateResponse {
+	private final Long id;
+}
