+++ src/main/java/com/sumin/paypay/repository/AccountRepositoryCustom.java
@@ -1,0 +1,8 @@
+package com.sumin.paypay.repository;
+
+import com.sumin.paypay.entity.Account;
+
+public interface AccountRepositoryCustom {
+
+	Account findByIdWithLock(Long accountId);
+}
