+++ src/main/java/com/sumin/paypay/repository/TransactionRepository.java
@@ -1,0 +1,8 @@
+package com.sumin.paypay.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.sumin.paypay.entity.Transaction;
+
+public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+}
