+++ src/main/java/com/sumin/paypay/type/TransactionType.java
@@ -1,0 +1,8 @@
+package com.sumin.paypay.type;
+
+import lombok.Getter;
+
+@Getter
+public enum TransactionType {
+
+}
