--- src/main/java/sumin/com/paypay/PayPayApplication.java
@@ -1,13 +1,0 @@
-package sumin.com.paypay;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PayPayApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PayPayApplication.class, args);
-	}
-
-}
