--- src/main/java/sumin/com/paypay/entity/Account.java
@@ -1,66 +1,0 @@
-package sumin.com.paypay.entity;
-
-import java.math.BigDecimal;
-import java.time.Instant;
-
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@NoArgsConstructor
-@Entity
-@Table(name = "account")
-public class Account {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "id")
-	private Long id;
-
-	@Column(name = "user_id")
-	private Long userId;
-
-	@Column(name = "account_number")
-	private String accountNumber;
-
-	@Column(name = "balance", nullable = false, precision = 19, scale = 2)
-	private BigDecimal balance;
-
-	@Column(name = "transaction_at")
-	private Instant transactionAt;
-
-	@Column(name = "created_at", nullable = false)
-	@CreationTimestamp
-	private Instant createdAt;
-
-	@Column(name = "updated_at", nullable = false)
-	@UpdateTimestamp
-	private Instant updatedAt;
-
-	@Column(name = "deleted_at")
-	private Instant deletedAt;
-
-	public static Account of(Long userId, String accountNumber) {
-		Account account = new Account();
-
-		account.userId = userId;
-		account.accountNumber = accountNumber;
-		account.balance = BigDecimal.ZERO;
-		account.transactionAt = null;
-		account.deletedAt = null;
-
-		return account;
-	}
-
-}
