--- src/main/java/sumin/com/paypay/entity/Transaction.java
@@ -1,61 +1,0 @@
-package sumin.com.paypay.entity;
-
-import java.math.BigDecimal;
-import java.time.Instant;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import sumin.com.paypay.type.TransactionType;
-
-@Getter
-@NoArgsConstructor
-@Entity
-@Table(name = "transaction")
-public class Transaction {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "id")
-	private Long id;
-
-	@Column(name = "from_account_id", nullable = false)
-	private Long fromAccountId;
-
-	@Column(name = "to_account_id")
-	private Long toAccountId;
-
-	@Column(name = "amount", nullable = false, precision = 19, scale = 2)
-	private BigDecimal amount;
-
-	@Column(name = "currency_type", nullable = false)
-	private String currencyType;
-
-	@Column(name = "bank_code")
-	private String bankCode;
-
-	@Column(name = "fee", precision = 8, scale = 2)
-	private BigDecimal fee;
-
-	@Size(max = 50)
-	@Enumerated(EnumType.STRING)
-	@Column(name = "transaction_type", nullable = false)
-	private TransactionType transactionType;
-
-	@Column(name = "created_at", nullable = false)
-	private Instant createdAt;
-
-	@Column(name = "updated_at", nullable = false)
-	private Instant updatedAt;
-
-}
