--- src/main/java/sumin/com/paypay/entity/User.java
@@ -1,47 +1,0 @@
-package sumin.com.paypay.entity;
-
-import java.time.Instant;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@Entity
-@Builder
-@Table(name = "user")
-@AllArgsConstructor
-public class User {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "id")
-	private Long id;
-
-	@Column(name = "name", nullable = false)
-	private String name;
-
-	@Column(name = "email", nullable = false)
-	private String email;
-
-	@Column(name = "pwd", nullable = false)
-	private String pwd;
-
-	@Column(name = "active", nullable = false)
-	private Boolean active;
-
-	@Column(name = "created_at", nullable = false)
-	private Instant createdAt;
-
-	@Column(name = "updated_at", nullable = false)
-	private Instant updatedAt;
-
-
-}
