--- src/main/java/sumin/com/paypay/exception/CommonException.java
@@ -1,12 +1,0 @@
-package sumin.com.paypay.exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class CommonException extends RuntimeException{
-
-	private final ResultCode resultCode;
-	private final String message;
-}
