--- src/main/java/sumin/com/paypay/exception/ResultCode.java
@@ -1,10 +1,0 @@
-package sumin.com.paypay.exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface ResultCode {
-
-	HttpStatus getStatus();
-
-	String getMessage();
-}
