--- src/main/java/sumin/com/paypay/model/response/AccountCreateResponse.java
@@ -1,10 +1,0 @@
-package sumin.com.paypay.model.response;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class AccountCreateResponse {
-	private final Long id;
-}
