--- src/main/java/sumin/com/paypay/repository/TransactionRepository.java
@@ -1,8 +1,0 @@
-package sumin.com.paypay.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import sumin.com.paypay.entity.Transaction;
-
-public interface TransactionRepository extends JpaRepository<Transaction, Long> {
-}
