--- src/main/java/sumin/com/paypay/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package sumin.com.paypay.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import sumin.com.paypay.entity.User;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
