--- src/main/java/sumin/com/paypay/type/TransactionType.java
@@ -1,8 +1,0 @@
-package sumin.com.paypay.type;
-
-import lombok.Getter;
-
-@Getter
-public enum TransactionType {
-
-}
