+++ src/test/java/com/sumin/paypay/PayPayApplicationTests.java
@@ -1,0 +1,13 @@
+package com.sumin.paypay;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PayPayApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
