--- src/test/java/sumin/com/paypay/PayPayApplicationTests.java
@@ -1,13 +1,0 @@
-package sumin.com.paypay;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PayPayApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
